pyproject.toml
[tool.poetry]
name = "dance_aggregator"
version = "0.1.0"
description = ""
authors = ["Augustus Ostow <ostowster@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
bs4 = "^0.0.1"
requests = "^2.22"
google-api-python-client = "^1.7"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.0"
pytz = "^2019.1"
lxml = "^4.3"
pandas = "^0.25.0"
pyperclip = "^1.7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
ipykernel = "^5.1"
autopep8 = "^1.4"
flake8 = "^3.7"
mypy = "^0.720.0"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
ipdb = "^0.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
